backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
jinja2
